Initial skeleton code for C++ Implementation
* Add libcask native dll project to sln
* Add stub implementation and P/Invoke interop to it to tests
* Revert back to CRLF on Windows, there are too many issues with VS putting snippets with CRLF into files and making them have mixed line endings
* Use MSBuild magic so that we
   1. Have a good experience in VS on Windows for C++
   2. Don't regress the C# dev experience x-plat for C#

This currently works only on Windows x64 and only with full VS, but everything is set up there to the point where you can drop a breakpoint in .cpp and hit it from the Test Explorer.

More work will be needed to match this experience for linux, mac, or even VS Code on Windows.