Fill more coverage holes. Remove test file code coverage exclusion attributes (#42)
This change replaces actual thought and discernment about what tests to
author by simply observing what's uncovered and closing those gaps. That
doesn't mean we shouldn't go back and revisit the tests for clarity and
consistent, i.e., to ensure that some code paths aren't being
opportunistically covered rather than by explicit, intentional tests.

But I think it's important and useful to first ensure that everything is
covered, so that we have that preliminary confidence. And, in fact, and
as usual, this exercise of closing all the little open places was
productive a bug fix.

Changes include:

- I've dropped the visibility of `Cask.ComputeExpiryChars` to private as
this data is proposed to become optional (and removed from the official
API). This prompted me to drop a `ThrowIf` protection to a debug assert,
as we now have complete control over all callers.
- `Helpers.TryByteToKind` now only validates the three legal cask key
kinds, rather than permitting any value that didn't clobber the reserved
padding in its relevant byte.
- Updated some comments.
- Mostly just added coverage. We now have basically 100% code coverage
in the API. There are some partial misses related to missed conditions
in a Debug.Assert (not relevant to production scenario). The only other
remaining gap is in a helper that's slated to get zapped when we
eliminate the explicit encoding of the sensitive component size.