Target netstandard2.0, remove System.IO.Hashing dependency
The library compiles fine for netstandard2.0 now, so use that in place of net472. Benchmarks and tests still use net472 to exercise the netstandard2.0 compilation. This aligns with https://github.com/security-utilities, which will make it easier to move code between these repos.

Also remove the System.IO.Hashing that we no longer use.